requirements.txt
aiohttp
beautifulsoup4
dnspython
aiofiles
